fix race conditions during parent<->child association
We need parent<->child association to inherit sandboxing data and establishing
this association is complicated in that there is no guarantee whether we will
receive PTRACE_EVENT_{FORK,VFORK,CLONE} - after which we can figure out the
child process ID using trace_geteventmsg() - or SIGSTOP of the new-born child
first.  Suspending the child until we receive one of the aforementioned ptrace
events is racy because the parent may be killed with SIGKILL beforehand,
leaving us with a child which is suspended forever...

Not suspending the child and waiting for the ptrace event gets problematic
because when we resume the new-born child in question, the ptrace event may
come too late (the child may have already entered a system call by then and we
need the sandboxing data to check for access).

These problems are caused due to the way sydbox wait()'s for events from
all processes simultaneously. To prevent races, add a `pidwait' variable
to sydbox_t and change this to the forking process ID on the entry of
fork(), vfork() and clone() system calls. This ensures we are going to
receive the aforementioned ptrace events before child creation.

However, there is a certain chance that the tracee in question may be
killed by SIGKILL before she returns from fork() but after the child is
created. In this rare case inheritance is not possible so use
the global (unmodified) sandboxing data.