config: simple configuration file naming scheme
The filename extension of a configuration file *must* start with ".syd-"
followed by the magic command API version. For sydbox-1 the magic
command API version starts from number 1 (which is the loneliest number
that you'll ever know). The current magic command API version is 1.

For profiles both -c @profile and -c @profile.syd-1 is supported. The
former defaults to the current profile.