aclq: correct EAFNOSUPPORT logic in acl_append_sockmatch
sockmatch_parse returns 0 and sets errno to EAFNOSUPPORT when sydbox
does not have IPv6 support and a IPv6 address is specified as argument.

acl_append_sockmatch incorrectly returned error in this case. Make
acl_append_sockmatch return 0 and set errno to EAFNOSUPPORT in case of
failure.

magic_errno function is renamed to magic_check_call which now checks
whether errno is set to EAFNOSUPPORT and returns MAGIC_RET_NOT_SUPPORTED
in this case. To return error is up to the caller.