optimise empty_dir heavily using getdents64()
With this very important optimisation, sydbox uses getdents64() system
call directly on Linux to read only 2 entries instead of a lot more
(800+ on the test machine). 2 entries are enough (think "." and "..") to
determine whether a directory is empty.

This is a huge leap forward making sydbox faster than sound.

Below is a very detailed statistical observation of the results of this
change.

Before:
% c99 -DSYDBOX_NO_GETDENTS empty.c
% strace -T -e getdents64,getdents ./a.out kingbee.d
getdents(3, /* 820 entries */, 32768)   = 32768 <0.000865>
+++ exited with 217 +++

After:
% c99 empty.c
% strace -T -e getdents64,getdents ./a.out kingbee.d
getdents64(3, /* 2 entries */, 64)      = 48 <0.000098>
+++ exited with 217 +++

This is the first of a series of optimisations aiming to eventually make
sydbox-1 speed catch the speed of light. (Next step is to make sydbox
faster for infinite loops.)