start working on the dumper
Logging was obviously a bad idea which is often not really useful when
debugging tracing issues which are hard to tackle. (Reading files of
+500M has never been fun :)

This commit is the start of the JSON event dumper which will be enabled
upon a configure switch (--enable-debug) and writes events of various
types in JSON format to a named pipe specified by the SHOEBOX variable.

A proof-of-concept script (shoebox.py) is written in Python which reads
from the named pipe and outputs the information to a file using bzip2
compression. It has primitive abilities to query the compressed json
file afterwards which are yet to be improved.

What's left to be done:
0. Go out to meet your girlfriend.
1. Write dump handlers for more events.
2. When (1) is done ditch logging entirely.
3. Report access violations only.
4. Improve shoebox.py query capabilities.
5. Use the dumper to understand and fix the sbcl hang.
6. Tell pipping he owes alip a poem when {1..5} is done.
7. Use shoebox.py to test sydbox internals more thoroughly.

Note, this commit also involves my initial attempts to figure out the
sbcl bug which drove me to write this dumper in the first place. I will
split these changes from this commit eventually.